+++ sdk/redisenterprise/arm-redisenterprisecache/src/models/mappers.ts
@@ -282,6 +282,129 @@ export const Sku: coreClient.CompositeMapper = {
   }
 };
 
+export const ManagedServiceIdentity: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ManagedServiceIdentity",
+    modelProperties: {
+      principalId: {
+        serializedName: "principalId",
+        readOnly: true,
+        type: {
+          name: "Uuid"
+        }
+      },
+      tenantId: {
+        serializedName: "tenantId",
+        readOnly: true,
+        type: {
+          name: "Uuid"
+        }
+      },
+      type: {
+        serializedName: "type",
+        required: true,
+        type: {
+          name: "String"
+        }
+      },
+      userAssignedIdentities: {
+        serializedName: "userAssignedIdentities",
+        type: {
+          name: "Dictionary",
+          value: {
+            type: { name: "Composite", className: "UserAssignedIdentity" }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const UserAssignedIdentity: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "UserAssignedIdentity",
+    modelProperties: {
+      principalId: {
+        serializedName: "principalId",
+        readOnly: true,
+        type: {
+          name: "Uuid"
+        }
+      },
+      clientId: {
+        serializedName: "clientId",
+        readOnly: true,
+        type: {
+          name: "Uuid"
+        }
+      }
+    }
+  }
+};
+
+export const ClusterPropertiesEncryption: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ClusterPropertiesEncryption",
+    modelProperties: {
+      customerManagedKeyEncryption: {
+        serializedName: "customerManagedKeyEncryption",
+        type: {
+          name: "Composite",
+          className: "ClusterPropertiesEncryptionCustomerManagedKeyEncryption"
+        }
+      }
+    }
+  }
+};
+
+export const ClusterPropertiesEncryptionCustomerManagedKeyEncryption: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "ClusterPropertiesEncryptionCustomerManagedKeyEncryption",
+    modelProperties: {
+      keyEncryptionKeyIdentity: {
+        serializedName: "keyEncryptionKeyIdentity",
+        type: {
+          name: "Composite",
+          className:
+            "ClusterPropertiesEncryptionCustomerManagedKeyEncryptionKeyIdentity"
+        }
+      },
+      keyEncryptionKeyUrl: {
+        serializedName: "keyEncryptionKeyUrl",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
+export const ClusterPropertiesEncryptionCustomerManagedKeyEncryptionKeyIdentity: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className:
+      "ClusterPropertiesEncryptionCustomerManagedKeyEncryptionKeyIdentity",
+    modelProperties: {
+      userAssignedIdentityResourceId: {
+        serializedName: "userAssignedIdentityResourceId",
+        type: {
+          name: "String"
+        }
+      },
+      identityType: {
+        serializedName: "identityType",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const PrivateEndpoint: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -350,6 +473,58 @@ export const Resource: coreClient.CompositeMapper = {
         type: {
           name: "String"
         }
+      },
+      systemData: {
+        serializedName: "systemData",
+        type: {
+          name: "Composite",
+          className: "SystemData"
+        }
+      }
+    }
+  }
+};
+
+export const SystemData: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "SystemData",
+    modelProperties: {
+      createdBy: {
+        serializedName: "createdBy",
+        type: {
+          name: "String"
+        }
+      },
+      createdByType: {
+        serializedName: "createdByType",
+        type: {
+          name: "String"
+        }
+      },
+      createdAt: {
+        serializedName: "createdAt",
+        type: {
+          name: "DateTime"
+        }
+      },
+      lastModifiedBy: {
+        serializedName: "lastModifiedBy",
+        type: {
+          name: "String"
+        }
+      },
+      lastModifiedByType: {
+        serializedName: "lastModifiedByType",
+        type: {
+          name: "String"
+        }
+      },
+      lastModifiedAt: {
+        serializedName: "lastModifiedAt",
+        type: {
+          name: "DateTime"
+        }
       }
     }
   }
@@ -367,6 +542,13 @@ export const ClusterUpdate: coreClient.CompositeMapper = {
           className: "Sku"
         }
       },
+      identity: {
+        serializedName: "identity",
+        type: {
+          name: "Composite",
+          className: "ManagedServiceIdentity"
+        }
+      },
       tags: {
         serializedName: "tags",
         type: {
@@ -378,6 +560,13 @@ export const ClusterUpdate: coreClient.CompositeMapper = {
         serializedName: "properties.minimumTlsVersion",
         type: {
           name: "String"
+        }
+      },
+      encryption: {
+        serializedName: "properties.encryption",
+        type: {
+          name: "Composite",
+          className: "ClusterPropertiesEncryption"
         }
       },
       hostName: {
@@ -805,6 +994,139 @@ export const ForceUnlinkParameters: coreClient.CompositeMapper = {
   }
 };
 
+export const FlushParameters: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "FlushParameters",
+    modelProperties: {
+      ids: {
+        serializedName: "ids",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "String"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const RegionSkuDetails: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "RegionSkuDetails",
+    modelProperties: {
+      value: {
+        serializedName: "value",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "RegionSkuDetail"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const RegionSkuDetail: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "RegionSkuDetail",
+    modelProperties: {
+      resourceType: {
+        serializedName: "resourceType",
+        type: {
+          name: "String"
+        }
+      },
+      locationInfo: {
+        serializedName: "locationInfo",
+        type: {
+          name: "Composite",
+          className: "LocationInfo"
+        }
+      },
+      skuDetails: {
+        serializedName: "skuDetails",
+        type: {
+          name: "Composite",
+          className: "SkuDetail"
+        }
+      }
+    }
+  }
+};
+
+export const LocationInfo: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "LocationInfo",
+    modelProperties: {
+      location: {
+        serializedName: "location",
+        type: {
+          name: "String"
+        }
+      },
+      capabilities: {
+        serializedName: "capabilities",
+        type: {
+          name: "Sequence",
+          element: {
+            type: {
+              name: "Composite",
+              className: "Capability"
+            }
+          }
+        }
+      }
+    }
+  }
+};
+
+export const Capability: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "Capability",
+    modelProperties: {
+      name: {
+        serializedName: "name",
+        type: {
+          name: "String"
+        }
+      },
+      value: {
+        serializedName: "value",
+        type: {
+          name: "Boolean"
+        }
+      }
+    }
+  }
+};
+
+export const SkuDetail: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "SkuDetail",
+    modelProperties: {
+      name: {
+        serializedName: "name",
+        type: {
+          name: "String"
+        }
+      }
+    }
+  }
+};
+
 export const PrivateEndpointConnection: coreClient.CompositeMapper = {
   type: {
     name: "Composite",
@@ -934,10 +1256,24 @@ export const Cluster: coreClient.CompositeMapper = {
           }
         }
       },
+      identity: {
+        serializedName: "identity",
+        type: {
+          name: "Composite",
+          className: "ManagedServiceIdentity"
+        }
+      },
       minimumTlsVersion: {
         serializedName: "properties.minimumTlsVersion",
         type: {
           name: "String"
+        }
+      },
+      encryption: {
+        serializedName: "properties.encryption",
+        type: {
+          name: "Composite",
+          className: "ClusterPropertiesEncryption"
         }
       },
       hostName: {
@@ -1053,6 +1389,27 @@ export const Database: coreClient.CompositeMapper = {
         type: {
           name: "Composite",
           className: "DatabasePropertiesGeoReplication"
+        }
+      }
+    }
+  }
+};
+
+export const DatabasesFlushHeaders: coreClient.CompositeMapper = {
+  type: {
+    name: "Composite",
+    className: "DatabasesFlushHeaders",
+    modelProperties: {
+      location: {
+        serializedName: "location",
+        type: {
+          name: "String"
+        }
+      },
+      azureAsyncOperation: {
+        serializedName: "azure-asyncoperation",
+        type: {
+          name: "String"
         }
       }
     }