+++ sdk/redisenterprise/arm-redisenterprisecache/src/operationsInterfaces/index.ts
@@ -12,3 +12,4 @@ export * from "./redisEnterprise";
 export * from "./databases";
 export * from "./privateEndpointConnections";
 export * from "./privateLinkResources";
+export * from "./skus";