+++ sdk/redisenterprise/arm-redisenterprisecache/CHANGELOG.md
@@ -1,15 +1,41 @@
 # Release History
+    
+## 2.3.0-beta.1 (2023-03-27)
+    
+**Features**
 
-## 2.2.1 (Unreleased)
-
-### Features Added
-
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Added operation group Skus
+  - Added operation Databases.beginFlush
+  - Added operation Databases.beginFlushAndWait
+  - Added Interface Capability
+  - Added Interface ClusterPropertiesEncryption
+  - Added Interface ClusterPropertiesEncryptionCustomerManagedKeyEncryption
+  - Added Interface ClusterPropertiesEncryptionCustomerManagedKeyEncryptionKeyIdentity
+  - Added Interface DatabasesFlushHeaders
+  - Added Interface DatabasesFlushOptionalParams
+  - Added Interface FlushParameters
+  - Added Interface LocationInfo
+  - Added Interface ManagedServiceIdentity
+  - Added Interface RegionSkuDetail
+  - Added Interface RegionSkuDetails
+  - Added Interface SkuDetail
+  - Added Interface SkusListOptionalParams
+  - Added Interface SystemData
+  - Added Interface UserAssignedIdentity
+  - Added Type Alias CmkIdentityType
+  - Added Type Alias CreatedByType
+  - Added Type Alias ManagedServiceIdentityType
+  - Added Type Alias SkusListResponse
+  - Interface Cluster has a new optional parameter encryption
+  - Interface Cluster has a new optional parameter identity
+  - Interface ClusterUpdate has a new optional parameter encryption
+  - Interface ClusterUpdate has a new optional parameter identity
+  - Interface Resource has a new optional parameter systemData
+  - Added Enum KnownCmkIdentityType
+  - Added Enum KnownCreatedByType
+  - Added Enum KnownManagedServiceIdentityType
+    
+    
 ## 2.2.0 (2022-12-07)
     
 **Features**